Create a new PSFFile class for parsing and writing PSF files, which allows me to offload the write_psf function from CharmmPsfFile to Structure and improve automatic file type detection.
