+++ chemistry/charmm/psf.py
@@ -13,8 +13,6 @@ from chemistry import (Bond, Angle, Dihedral, Improper, AcceptorDonor, Group,
                        Cmap, UreyBradley, NoUreyBradley, Structure, Atom)
 from chemistry.exceptions import (CharmmPSFError, MoleculeError,
                 CharmmPSFWarning, MissingParameter, CharmmPsfEOF)
-from chemistry.formats import io
-from chemistry.formats.registry import FileFormatType
 from chemistry.structure import needs_openmm, app, mm
 from chemistry import unit as u
 from math import sqrt
@@ -83,31 +81,6 @@ class CharmmPsfFile(Structure):
     >>> len(cs.bonds)
     32
     """
-    __metaclass__ = FileFormatType
-
-    #===================================================
-
-    @staticmethod
-    def id_format(filename):
-        """ Identifies the file type as a CHARMM PSF file
-
-        Parameters
-        ----------
-        filename : str
-            Name of the file to check format for
-
-        Returns
-        -------
-        is_fmt : bool
-            True if it is a CHARMM or Xplor-style PSF file
-        """
-        f = io.genopen(filename, 'r')
-        line = f.readline().decode()
-        f.close()
-        return line.strip().startswith('PSF')
-
-    #===================================================
-
     @staticmethod
     def _convert(string, type, message):
         """
@@ -357,187 +330,6 @@ class CharmmPsfFile(Structure):
 
     #===================================================
 
-    def write_psf(self, dest, vmd=False):
-        """
-        Writes a PSF file from the stored molecule
-
-        Parameters
-        ----------
-        dest : str or file-like
-            The place to write the output PSF file.  If it has a "write"
-            attribute, it will be used to print the PSF file. Otherwise, it will
-            be treated like a string and a file will be opened, printed, then
-            closed
-        vmd : bool 
-            If True, it will write out a PSF in the format that VMD prints it in
-            (i.e., no NUMLP/NUMLPH or MOLNT sections)
-
-        Examples
-        --------
-        >>> cs = CharmmPsfFile('testfiles/test.psf')
-        >>> cs.write_psf('testfiles/test2.psf')
-        """
-        # See if this is an extended format
-        try:
-            ext = 'EXT' in self.flags
-        except AttributeError:
-            ext = True
-        own_handle = False
-        # Index the atoms and residues
-        if not hasattr(dest, 'write'):
-            own_handle = True
-            dest = open(dest, 'w')
-
-        # Assign the formats we need to write with
-        if ext:
-            atmfmt1 = ('%10d %-8s %-8i %-8s %-8s %4d %10.6f %13.4f' + 11*' ')
-            atmfmt2 = ('%10d %-8s %-8i %-8s %-8s %-4s %10.6f %13.4f' + 11*' ')
-            intfmt = '%10d' # For pointers
-        else:
-            atmfmt1 = ('%8d %-4s %-4i %-4s %-4s %4d %10.6f %13.4f' + 11*' ')
-            atmfmt2 = ('%8d %-4s %-4i %-4s %-4s %-4s %10.6f %13.4f' + 11*' ')
-            intfmt = '%8d' # For pointers
-
-        # Now print the header then the title
-        dest.write('PSF ')
-        if hasattr(self, 'flags'):
-            dest.write(' '.join(self.flags))
-        else:
-            dest.write('EXT') # EXT is always active
-        dest.write('\n\n')
-        dest.write(intfmt % len(self.title) + ' !NTITLE\n')
-        dest.write('\n'.join(self.title) + '\n\n')
-        # Now time for the atoms
-        dest.write(intfmt % len(self.atoms) + ' !NATOM\n')
-        # atmfmt1 is for CHARMM format (i.e., atom types are integers)
-        # atmfmt is for XPLOR format (i.e., atom types are strings)
-        for i, atom in enumerate(self.atoms):
-            if isinstance(atom.type, str):
-                fmt = atmfmt2
-            else:
-                fmt = atmfmt1
-            if hasattr(atom, 'segid'):
-                segid = atom.segid
-            else:
-                segid = 'SYS'
-            atmstr = fmt % (i+1, segid, atom.residue.number,
-                            atom.residue.name, atom.name, atom.type,
-                            atom.charge, atom.mass)
-            if hasattr(atom, 'props'):
-                dest.write(atmstr + '   '.join(atom.props) + '\n')
-            else:
-                dest.write('\n')
-        dest.write('\n')
-        # Bonds
-        dest.write(intfmt % len(self.bonds) + ' !NBOND: bonds\n')
-        for i, bond in enumerate(self.bonds):
-            dest.write((intfmt*2) % (bond.atom1.idx+1, bond.atom2.idx+1))
-            if i % 4 == 3: # Write 4 bonds per line
-                dest.write('\n')
-        # See if we need to terminate
-        if len(self.bonds) % 4 != 0 or len(self.bonds) == 0:
-            dest.write('\n')
-        dest.write('\n')
-        # Angles
-        dest.write(intfmt % len(self.angles) + ' !NTHETA: angles\n')
-        for i, angle in enumerate(self.angles):
-            dest.write((intfmt*3) % (angle.atom1.idx+1, angle.atom2.idx+1,
-                                     angle.atom3.idx+1)
-            )
-            if i % 3 == 2: # Write 3 angles per line
-                dest.write('\n')
-        # See if we need to terminate
-        if len(self.angles) % 3 != 0 or len(self.angles) == 0:
-            dest.write('\n')
-        dest.write('\n')
-        # Dihedrals
-        dest.write(intfmt % len(self.dihedrals) + ' !NPHI: dihedrals\n')
-        for i, dih in enumerate(self.dihedrals):
-            dest.write((intfmt*4) % (dih.atom1.idx+1, dih.atom2.idx+1,
-                                     dih.atom3.idx+1, dih.atom4.idx+1)
-            )
-            if i % 2 == 1: # Write 2 dihedrals per line
-                dest.write('\n')
-        # See if we need to terminate
-        if len(self.dihedrals) % 2 != 0 or len(self.dihedrals) == 0:
-            dest.write('\n')
-        dest.write('\n')
-        # Impropers
-        dest.write(intfmt % len(self.impropers) + ' !NIMPHI: impropers\n')
-        for i, imp in enumerate(self.impropers):
-            dest.write((intfmt*4) % (imp.atom1.idx+1, imp.atom2.idx+1,
-                                     imp.atom3.idx+1, imp.atom4.idx+1)
-            )
-            if i % 2 == 1: # Write 2 dihedrals per line
-                dest.write('\n')
-        # See if we need to terminate
-        if len(self.impropers) % 2 != 0 or len(self.impropers) == 0:
-            dest.write('\n')
-        dest.write('\n')
-        # Donor section
-        dest.write(intfmt % len(self.donors) + ' !NDON: donors\n')
-        for i, don in enumerate(self.donors):
-            dest.write((intfmt*2) % (don.atom1.idx+1, don.atom2.idx+1))
-            if i % 4 == 3: # 4 donors per line
-                dest.write('\n')
-        if len(self.donors) % 4 != 0 or len(self.donors) == 0:
-            dest.write('\n')
-        dest.write('\n')
-        # Acceptor section
-        dest.write(intfmt % len(self.acceptors) + ' !NACC: acceptors\n')
-        for i, acc in enumerate(self.acceptors):
-            dest.write((intfmt*2) % (acc.atom1.idx+1, acc.atom2.idx+1))
-            if i % 4 == 3: # 4 donors per line
-                dest.write('\n')
-        if len(self.acceptors) % 4 != 0 or len(self.acceptors) == 0:
-            dest.write('\n')
-        dest.write('\n')
-        # NNB section ??
-        dest.write(intfmt % 0 + ' !NNB\n\n')
-        for i in xrange(len(self.atoms)):
-            dest.write(intfmt % 0)
-            if i % 8 == 7: # Write 8 0's per line
-                dest.write('\n')
-        if len(self.atoms) % 8 != 0: dest.write('\n')
-        dest.write('\n')
-        # Group section
-        dest.write((intfmt*2) % (len(self.groups), self.groups.nst2))
-        dest.write(' !NGRP NST2\n')
-        for i, gp in enumerate(self.groups):
-            dest.write((intfmt*3) % (gp.bs, gp.type, gp.move))
-            if i % 3 == 2: dest.write('\n')
-        if len(self.groups) % 3 != 0 or len(self.groups) == 0:
-            dest.write('\n')
-        dest.write('\n')
-        # The next two sections are never found in VMD prmtops...
-        if not vmd:
-            # Molecule section; first set molecularity
-            set_molecules(self.atoms)
-            mollist = [a.marked for a in self.atoms]
-            dest.write(intfmt % max(mollist) + ' !MOLNT\n')
-            for i, atom in enumerate(self.atoms):
-                dest.write(intfmt % atom.marked)
-                if i % 8 == 7: dest.write('\n')
-            if len(self.atoms) % 8 != 0: dest.write('\n')
-            dest.write('\n')
-            # NUMLP/NUMLPH section
-            dest.write((intfmt*2) % (0, 0) + ' !NUMLP NUMLPH\n')
-            dest.write('\n')
-        # CMAP section
-        dest.write(intfmt % len(self.cmaps) + ' !NCRTERM: cross-terms\n')
-        for i, cmap in enumerate(self.cmaps):
-            dest.write((intfmt*8) % (cmap.atom1.idx+1, cmap.atom2.idx+1,
-                                     cmap.atom3.idx+1, cmap.atom4.idx+1,
-                                     cmap.atom2.idx+1, cmap.atom3.idx+1,
-                                     cmap.atom4.idx+1, cmap.atom5.idx+1)
-            )
-            dest.write('\n')
-        # Done!
-        # If we opened our own handle, close it
-        if own_handle:
-            dest.close()
-
-    #===================================================
 
     @needs_openmm
     def createSystem(self, params=None, *args, **kwargs):
